1:53
Implementation of details screen with the desired design and server support.
Implementation of logout.

+++ app/src/main/java/ru/tsu/ibrahimfall/notforget/model/items/task/NewTask.kt
@@ -7,7 +7,7 @@ data class NewTask(
     var title: String? = null,
     var description: String? = null,
     var done: Int = 0,
-    val deadline: Int = 0,
+    val deadline: Long = 0,
     @SerializedName(CATEGORY_SERIALIZATION)
     var category: Int? = null,
     @SerializedName(PRIORITY_SERIALIZATION)
